Refactor fetchAllEvents to the workflow-history-layout. Fix timeline issue with json. Add loading to timeline
