+++ src/lib/pages/workflow-history-json.svelte
@@ -1,28 +1,12 @@
 <script lang="ts">
-  import { page } from '$app/stores';
-
   import WorkflowJsonNavigator from '$lib/components/workflow/workflow-json-navigator.svelte';
   import ToggleSwitch from '$lib/holocene/toggle-switch.svelte';
-  import { fetchRawEvents } from '$lib/services/events-service';
-  import { decodeURIForSvelte } from '$lib/utilities/encode-uri';
-
-  const { namespace, workflow: workflowId, run: runId } = $page.params;
+  import { fullEventHistory } from '$lib/stores/events';
 
   let decodeEventHistory = true;
-  let events = [];
-
-  const fetchEvents = async () => {
-    events = await fetchRawEvents({
-      namespace: decodeURIForSvelte(namespace),
-      workflowId: decodeURIForSvelte(workflowId),
-      runId: decodeURIForSvelte(runId),
-      sort: 'ascending',
-    });
-  };
-  $: decodeEventHistory, fetchEvents();
 </script>
 
-<WorkflowJsonNavigator {events} {decodeEventHistory}>
+<WorkflowJsonNavigator events={$fullEventHistory} {decodeEventHistory}>
   <ToggleSwitch
     slot="decode"
     label="View decoded event history"
